fix\delete

+++ backend/controllers/cards.js
@@ -1,6 +1,7 @@
 const Card = require('../models/card.js');
 const { NotFoundError } = require('../utils/NotFoundError.js');
 const { InvalidError } = require('../utils/InvalidError.js');
+const { createDecipher } = require('crypto');
 
 const getCards = (req, res, next) => {
   Card.find({})
@@ -25,20 +26,34 @@ const postCard = (req, res, next) => {
 };
 
 const getCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.id)
-    .orFail(() => {
-      const error = new NotFoundError('Такой карточки не существует');
-      throw error;
-    })
-    .then((card) => res.status(200).send({ data: card }))
-    .catch((err) => {
-      if (err.kind === 'ObjectId') {
-        const error = new InvalidError('Такой карточки не существует');
-        next(error);
-      } else {
-        next(err);
-      }
-    });
+  Card.findById(req.params.id)
+  .orFail(() => {
+    const error = new NotFoundError('Такой карточки не существует');
+    throw error;
+  })
+  .then ((card) => {
+    if (card.owner.toString() === req.user.id) {
+      Card.findByIdAndRemove(req.params.id)
+      .then((card) => {
+        res.status(200).send({ data: card })
+      })
+      .catch((err) => {
+        if (err.kind === 'ObjectId') {
+          const error = new InvalidError('Такой карточки не существует');
+          next(error);
+        } else {
+          next(err);
+        }
+      });
+    }
+    else {
+      const error = new InvalidError('Это не ваша карточка');
+      next(error);
+    }
+  })
+  .catch((err) => {
+    next(err)
+  });
 };
 
 const putLike = (req, res, next) => {
